# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MrRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MrRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/WCY-dt/MrRSS",
            "email": "mail@ch3nyang.top"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.3.19"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/articles/cleanup": {
            "post": {
                "description": "Delete all articles and article contents (keeps feeds and settings)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Cleanup all articles",
                "responses": {
                    "200": {
                        "description": "Cleanup statistics (deleted, articles, contents, type)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/articles/cleanup-content": {
            "post": {
                "description": "Clear all cached article content (articles remain, only content cache is cleared)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Cleanup article content cache",
                "responses": {
                    "200": {
                        "description": "Cleanup result (success, entries_cleaned)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/articles/clear-read-later": {
            "post": {
                "description": "Remove all articles from the read-later list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Clear read-later list",
                "responses": {
                    "200": {
                        "description": "Read-later list cleared successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/content-cache-info": {
            "get": {
                "description": "Get statistics about the article content cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get article content cache info",
                "responses": {
                    "200": {
                        "description": "Cache info (cached_articles count)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/filter-counts": {
            "get": {
                "description": "Get per-feed counts for different filter types (unread, favorites, read_later, images)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get filter-specific feed counts",
                "responses": {
                    "200": {
                        "description": "Filter counts for all filter types",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/mark-all-read": {
            "post": {
                "description": "Mark all articles as read globally, by feed, or by category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Mark all articles as read",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Mark all as read for specific feed ID",
                        "name": "feed_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Mark all as read for specific category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles marked as read successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid feed_id)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/mark-relative": {
            "post": {
                "description": "Marks articles as read based on their published time relative to a reference article (above = newer, below = older)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Mark articles relative to reference article",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Reference article ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "above",
                            "below"
                        ],
                        "type": "string",
                        "description": "Direction: 'above' for newer articles, 'below' for older articles",
                        "name": "direction",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Optional: only mark articles from this feed",
                        "name": "feed_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Optional: only mark articles from this category",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Number of articles marked as read",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/articles/unread-counts": {
            "get": {
                "description": "Get total unread count and per-feed unread counts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get unread counts",
                "responses": {
                    "200": {
                        "description": "Unread counts (total + feed_counts map)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}